+++ lib/read_parsers.hh
@@ -38,13 +38,19 @@ Contact: khmer-project@idyll.org
 #ifndef READ_PARSERS_HH
 #define READ_PARSERS_HH
 
+#include <seqan/seq_io.h> // IWYU pragma: keep
+#include <seqan/sequence.h> // IWYU pragma: keep
+#include <seqan/stream.h> // IWYU pragma: keep
+
 #include <regex.h>
 #include <stddef.h>
 #include <stdint.h>
+#include <algorithm>
 #include <cstdlib>
 #include <iostream>
 #include <string>
 #include <utility>
+#include <memory>
 
 #include "khmer.hh"
 #include "khmer_exception.hh"
@@ -83,16 +89,19 @@ struct InvalidReadPair : public  khmer_value_exception {
         khmer_value_exception("Invalid read pair detected.") {}
 };
 
-struct Read {
+
+unsigned char _to_valid_dna(const unsigned char c);
+
+
+struct Read
+{
     std::string name;
     std::string description;
     std::string sequence;
     std::string quality;
     std::string cleaned_seq;
 
-    //Read() : name(), annotations(), sequence(), quality(), cleaned_seq() {}
-
-    inline void reset( )
+    inline void reset()
     {
         name.clear();
         description.clear();
@@ -101,61 +110,87 @@ struct Read {
         cleaned_seq.clear();
     }
 
-    void write_to(std::ostream&);
+    inline void write_fastx(std::ostream& output)
+    {
+        if (quality.length() != 0) {
+            output << "@" << name << '\n'
+                   << sequence << '\n'
+                   << "+" << '\n'
+                   << quality << '\n';
+        } else {
+            output << ">" << name << '\n'
+                   << sequence << '\n';
+        }
+    }
+
+    // Compute cleaned_seq from sequence. Call this after changing sequence.
+    inline void set_clean_seq()
+    {
+        cleaned_seq = std::string(sequence.size(), 0);
+        std::transform(sequence.begin(), sequence.end(), cleaned_seq.begin(),
+                       _to_valid_dna);
+    }
 };
 typedef std::pair<Read, Read> ReadPair;
 
-class IParser
+
+template<typename SeqIO>
+class ReadParser
 {
+protected:
+    std::unique_ptr<SeqIO> _parser;
+    regex_t _re_read_2_nosub;
+    regex_t _re_read_1;
+    regex_t _re_read_2;
+    void _init();
+
+    ReadPair _get_next_read_pair_in_ignore_mode();
+    ReadPair _get_next_read_pair_in_error_mode();
+    bool _is_valid_read_pair(
+        ReadPair &the_read_pair,
+        regmatch_t &match_1,
+        regmatch_t &match_2
+    );
+
 public:
     enum {
         PAIR_MODE_IGNORE_UNPAIRED,
         PAIR_MODE_ERROR_ON_UNPAIRED
     };
 
-    static IParser * const get_parser(
-        std::string const &ifile_name
-    );
+    explicit ReadParser(std::unique_ptr<SeqIO> pf);
+    explicit ReadParser(ReadParser& other);
+    virtual ~ReadParser();
 
-    IParser();
-    virtual ~IParser();
+    Read get_next_read();
+    ReadPair get_next_read_pair(uint8_t mode = PAIR_MODE_ERROR_ON_UNPAIRED);
 
-    virtual bool is_complete( ) = 0;
-    virtual Read get_next_read() = 0;
-    virtual ReadPair get_next_read_pair(uint8_t mode = PAIR_MODE_ERROR_ON_UNPAIRED);
-    size_t get_num_reads()
-    {
-        return _num_reads;
-    }
+    size_t get_num_reads();
+    bool is_complete();
+}; // class ReadParser
 
-protected:
+
+class FastxReader
+{
+private:
+    std::string _filename;
+    seqan::SequenceStream _stream;
+    uint32_t _spin_lock;
     size_t _num_reads;
     bool _have_qualities;
-    regex_t _re_read_2_nosub;
-    regex_t _re_read_1;
-    regex_t _re_read_2;
+    void _init();
 
-    ReadPair _get_next_read_pair_in_ignore_mode();
-    ReadPair _get_next_read_pair_in_error_mode();
-    bool _is_valid_read_pair(
-        ReadPair &the_read_pair, regmatch_t &match_1, regmatch_t &match_2
-    );
-
-}; // class IParser
-
-class FastxParser : public IParser
-{
 public:
-    explicit FastxParser(const char * filename);
-    ~FastxParser();
+    FastxReader();
+    FastxReader(const std::string& infile);
+    FastxReader(FastxReader& other);
+    ~FastxReader();
 
-    bool is_complete();
     Read get_next_read();
+    bool is_complete();
+    size_t get_num_reads();
+}; // class FastxReader
 
-private:
-    struct Handle;
-    Handle* _private;
-};
 
 inline PartitionID _parse_partition_id(std::string name)
 {
@@ -179,6 +214,16 @@ inline PartitionID _parse_partition_id(std::string name)
 
     return p;
 }
+
+// Alias for generic/templated ReadParser pointer
+template<typename T> using ReadParserPtr = std::unique_ptr<ReadParser<T>>;
+
+// Convenience function
+template<typename SeqIO>
+ReadParserPtr<SeqIO> get_parser(const std::string& filename);
+
+// Alias for instantiated ReadParsers
+typedef std::unique_ptr<ReadParser<FastxReader>> FastxParserPtr;
 
 } // namespace read_parsers